Refactored: Datadog::Runtime::Metrics --> Datadog::Core::Runtime::Metrics
