+++ integration/apps/rack/app/datadog.rb
@@ -4,12 +4,12 @@ require 'ddtrace'
 Datadog.configure do |c|
   c.service = 'acme-rack'
   c.diagnostics.debug = true if Datadog::DemoEnv.feature?('debug')
+  c.runtime_metrics.enabled = true if Datadog::DemoEnv.feature?('runtime_metrics')
 end
 
 if Datadog::DemoEnv.feature?('tracing')
   Datadog::Tracing.configure do |c|
     c.analytics.enabled = true if Datadog::DemoEnv.feature?('analytics')
-    c.runtime_metrics.enabled = true if Datadog::DemoEnv.feature?('runtime_metrics')
     c.instrument :rack
   end
 end
