+++ integration/apps/rspec/app/datadog.rb
@@ -3,11 +3,11 @@ require 'ddtrace'
 
 Datadog.configure do |c|
   c.diagnostics.debug = true if Datadog::DemoEnv.feature?('debug')
+  c.runtime_metrics.enabled = true if Datadog::DemoEnv.feature?('runtime_metrics')
 end
 
 Datadog::Tracing.configure do |c|
   c.analytics.enabled = true if Datadog::DemoEnv.feature?('analytics')
-  c.runtime_metrics.enabled = true if Datadog::DemoEnv.feature?('runtime_metrics')
 end
 
 Datadog::Profiling.configure do |c|
