+++ lib/ddtrace/configuration/components.rb
@@ -36,7 +36,7 @@ module Datadog
           options[:statsd] = settings.runtime_metrics.statsd unless settings.runtime_metrics.statsd.nil?
           options[:services] = [settings.service] unless settings.service.nil?
 
-          Datadog::Runtime::Metrics.new(**options)
+          Datadog::Core::Runtime::Metrics.new(**options)
         end
 
         def build_runtime_metrics_worker(settings)
