+++ spec/ddtrace/configuration/components_spec.rb
@@ -186,12 +186,12 @@ RSpec.describe Datadog::Configuration::Components do
 
     context 'given settings' do
       shared_examples_for 'new runtime metrics' do
-        let(:runtime_metrics) { instance_double(Datadog::Runtime::Metrics) }
+        let(:runtime_metrics) { instance_double(Datadog::Core::Runtime::Metrics) }
         let(:default_options) { { enabled: settings.runtime_metrics.enabled, services: [settings.service] } }
         let(:options) { {} }
 
         before do
-          expect(Datadog::Runtime::Metrics).to receive(:new)
+          expect(Datadog::Core::Runtime::Metrics).to receive(:new)
             .with(default_options.merge(options))
             .and_return(runtime_metrics)
         end
@@ -253,7 +253,7 @@ RSpec.describe Datadog::Configuration::Components do
     context 'given settings' do
       shared_examples_for 'new runtime metrics worker' do
         let(:runtime_metrics_worker) { instance_double(Datadog::Workers::RuntimeMetrics) }
-        let(:runtime_metrics) { instance_double(Datadog::Runtime::Metrics) }
+        let(:runtime_metrics) { instance_double(Datadog::Core::Runtime::Metrics) }
         let(:default_options) do
           {
             enabled: settings.runtime_metrics.enabled,
@@ -1123,7 +1123,7 @@ RSpec.describe Datadog::Configuration::Components do
         let(:tracer) { instance_double(Datadog::Tracer) }
         let(:profiler) { Datadog::Profiling.supported? ? instance_double(Datadog::Profiler) : nil }
         let(:runtime_metrics_worker) { instance_double(Datadog::Workers::RuntimeMetrics, metrics: runtime_metrics) }
-        let(:runtime_metrics) { instance_double(Datadog::Runtime::Metrics, statsd: statsd) }
+        let(:runtime_metrics) { instance_double(Datadog::Core::Runtime::Metrics, statsd: statsd) }
         let(:health_metrics) { instance_double(Datadog::Diagnostics::Health::Metrics, statsd: statsd) }
         let(:statsd) { instance_double(::Datadog::Statsd) }
