+++ spec/ddtrace/contrib/rack/configuration_spec.rb
@@ -80,7 +80,7 @@ RSpec.describe 'Rack integration configuration' do
         expect(queue_span.span_type).to eq('proxy')
         expect(queue_span.service).to eq(web_service_name)
         expect(queue_span.start_time.to_i).to eq(queue_time)
-        expect(queue_span.get_tag(Datadog::Ext::Runtime::TAG_LANG)).to be_nil
+        expect(queue_span.get_tag(Datadog::Core::Runtime::Ext::TAG_LANG)).to be_nil
         expect(queue_span.get_tag(Datadog::Ext::Metadata::TAG_PEER_SERVICE)).to eq(web_service_name)
         expect(queue_span.get_tag(Datadog::Ext::Metadata::TAG_COMPONENT))
           .to eq('rack')
